1.fix: 修复主页动态组件在控制台报错的问题 2.feat: 支持修改个人主页 3.refactor: 优化权限控制, 当登录过期或禁止访问资源时自动跳转到登录界面 4.refactor: 优化界面加载动画效果

+++ src/components/main/profile/Password.vue
@@ -62,7 +62,7 @@ export default {
   methods: {
     // 获取验证码
     async getCode () {
-      const { data: result } = await this.$http.get('/verify/update', {
+      const { data: result } = await this.$http.get('/auth/verify/update', {
         params: { id: window.sessionStorage.getItem('uid') }
       })
       if (result.status !== 200) {
@@ -72,16 +72,21 @@ export default {
     },
     // 修改密码
     async updatePassword () {
-      const { data: result } = await this.$http.put(
-        '/admin/' + window.sessionStorage.getItem('uid') + '/password',
-        this.updatePasswordForm
-      )
-      if (result.status !== 200) {
-        return this.$message.error(result.message)
-      }
-      this.$message.success(result.message)
-      window.sessionStorage.clear()
-      this.$router.push('/login')
+      this.$refs.updatePasswordFormRef.validate(async (valid) => {
+        // 表单验证通过
+        if (valid) {
+          const { data: result } = await this.$http.put(
+            '/admin/' + window.sessionStorage.getItem('uid') + '/password',
+            this.updatePasswordForm
+          )
+          if (result.status !== 200) {
+            return this.$message.error(result.message)
+          }
+          this.$message.success(result.message)
+          window.sessionStorage.clear()
+          this.$router.push('/login')
+        }
+      })
     },
     // 取消修改密码
     cancelUpdatePassword (refName) {
